Combine SwapAssetsUpdateRunner and SwappableAssetsRunner into SwapAssetsRunner
Merge two related runners into a single runner that:
- Sets swappable assets based on supported chains (always)
- Updates swap assets from API when version changes
